backend/requirements.txt
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.7.0
python-dotenv==1.0.1
flake8
flake8-isort
Pillow
tqdm